requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-multipart==0.0.6
gunicorn==21.2.0
google-cloud-speech==2.21.0
google-cloud-texttospeech==2.14.1
google-generativeai==0.3.0
python-dotenv==1.0.0